// source/frontend/src/models/api/generated/ApiGen_Base_BaseAudit.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDateTime } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_BaseConcurrent } from './ApiGen_Base_BaseConcurrent';

// LINK: @backend/apimodels/Models/Base/BaseAuditModel.cs
export interface ApiGen_Base_BaseAudit extends ApiGen_Base_BaseConcurrent {
  appCreateTimestamp: UtcIsoDateTime;
  appLastUpdateTimestamp: UtcIsoDateTime;
  appLastUpdateUserid: string | null;
  appCreateUserid: string | null;
  appLastUpdateUserGuid: string | null;
  appCreateUserGuid: string | null;
}